Changes made to routing before moving on to setting up first fetch request
