+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.css'
-import Header from './Header'
+import Header from './Components/Header'
 import { Route, Routes } from 'react-router'
